pyproject.toml
[project]
name = "venus-a-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "livekit-agents[openai,silero,speechmatics]==1.4.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "uvicorn[standard]>=0.46.0",
]
